refactor: 유저 관련 usecase 분리
