+++ src/main/java/leets/weeth/domain/user/application/usecase/UserUseCase.java
@@ -1,6 +1,8 @@
 package leets.weeth.domain.user.application.usecase;
 
+import jakarta.servlet.http.HttpServletRequest;
 import leets.weeth.domain.user.application.dto.request.UserRequestDto;
+import leets.weeth.global.auth.jwt.application.dto.JwtDto;
 
 import java.util.List;
 import java.util.Map;
@@ -17,27 +19,6 @@ public interface UserUseCase {
 
     void register(Register dto, Long userId);
 
-    Response find(Long userId);
+    JwtDto refresh(HttpServletRequest request);
 
-    Map<Integer, List<Response>> findAll();
-
-    Map<Integer, List<SummaryResponse>> findAllUser();
-
-    List<AdminResponse> findAllByAdmin();
-
-    UserResponse findUserDetails(Long userId);
-
-    void update(Update dto, Long userId);
-
-    void accept(Long userId);
-
-    void update(Long userId, String role);
-
-    void leave(Long userId);
-
-    void ban(Long userId);
-
-    void applyOB(Long userId, Integer cardinal);
-
-    void reset(Long userId);
 }
